fix: serialize policy bigints
